requirements.txt
# Core dependencies for the image match prompt optimizer
mlflow==3.10.1
google-genai==1.70.0
litellm==1.81.13
gepa==0.1.1
Pillow==12.1.0
pandas==2.3.3
requests==2.32.5